Many wording/functionality improvements on monoalphabetic substitution and analysis plugin for monoalphabetic substitution

+++ org.jcryptool.analysis.friedman/src/org/jcryptool/analysis/friedman/ui/MColor.java
@@ -92,9 +92,9 @@ public class MColor {
         alpha = Alpha;
 
         String hexStr = hex.toLowerCase();
-        r = (int) Integer.parseInt("" + hexStr.charAt(0) + hexStr.charAt(1), 16); //$NON-NLS-1$
-        g = (int) Integer.parseInt("" + hexStr.charAt(2) + hexStr.charAt(3), 16); //$NON-NLS-1$
-        b = (int) Integer.parseInt("" + hexStr.charAt(4) + hexStr.charAt(5), 16); //$NON-NLS-1$
+        r = Integer.parseInt("" + hexStr.charAt(0) + hexStr.charAt(1), 16); //$NON-NLS-1$
+        g = Integer.parseInt("" + hexStr.charAt(2) + hexStr.charAt(3), 16); //$NON-NLS-1$
+        b = Integer.parseInt("" + hexStr.charAt(4) + hexStr.charAt(5), 16); //$NON-NLS-1$
 
         myColor = new Color(Display.getDefault(), r, g, b);
     }
@@ -108,9 +108,9 @@ public class MColor {
         alpha = Alpha;
 
         String hexStr = hex.toLowerCase();
-        r = (int) Integer.parseInt("" + hexStr.charAt(0) + hexStr.charAt(1), 16); //$NON-NLS-1$
-        g = (int) Integer.parseInt("" + hexStr.charAt(2) + hexStr.charAt(3), 16); //$NON-NLS-1$
-        b = (int) Integer.parseInt("" + hexStr.charAt(4) + hexStr.charAt(5), 16); //$NON-NLS-1$
+        r = Integer.parseInt("" + hexStr.charAt(0) + hexStr.charAt(1), 16); //$NON-NLS-1$
+        g = Integer.parseInt("" + hexStr.charAt(2) + hexStr.charAt(3), 16); //$NON-NLS-1$
+        b = Integer.parseInt("" + hexStr.charAt(4) + hexStr.charAt(5), 16); //$NON-NLS-1$
 
         myColor = new Color(Display.getDefault(), r, g, b);
 
@@ -125,9 +125,9 @@ public class MColor {
         alpha = Alpha;
 
         String hexStr = intToHexStr(asInt);
-        r = (int) Integer.parseInt("" + hexStr.charAt(0) + hexStr.charAt(1), 16); //$NON-NLS-1$
-        g = (int) Integer.parseInt("" + hexStr.charAt(2) + hexStr.charAt(3), 16); //$NON-NLS-1$
-        b = (int) Integer.parseInt("" + hexStr.charAt(4) + hexStr.charAt(5), 16); //$NON-NLS-1$
+        r = Integer.parseInt("" + hexStr.charAt(0) + hexStr.charAt(1), 16); //$NON-NLS-1$
+        g = Integer.parseInt("" + hexStr.charAt(2) + hexStr.charAt(3), 16); //$NON-NLS-1$
+        b = Integer.parseInt("" + hexStr.charAt(4) + hexStr.charAt(5), 16); //$NON-NLS-1$
 
         myColor = new Color(Display.getDefault(), r, g, b);
     }
@@ -140,9 +140,9 @@ public class MColor {
         alpha = Alpha;
 
         String hexStr = intToHexStr(asInt);
-        r = (int) Integer.parseInt("" + hexStr.charAt(0) + hexStr.charAt(1), 16); //$NON-NLS-1$
-        g = (int) Integer.parseInt("" + hexStr.charAt(2) + hexStr.charAt(3), 16); //$NON-NLS-1$
-        b = (int) Integer.parseInt("" + hexStr.charAt(4) + hexStr.charAt(5), 16); //$NON-NLS-1$
+        r = Integer.parseInt("" + hexStr.charAt(0) + hexStr.charAt(1), 16); //$NON-NLS-1$
+        g = Integer.parseInt("" + hexStr.charAt(2) + hexStr.charAt(3), 16); //$NON-NLS-1$
+        b = Integer.parseInt("" + hexStr.charAt(4) + hexStr.charAt(5), 16); //$NON-NLS-1$
 
         myColor = new Color(Display.getDefault(), r, g, b);
